클래스 이름 변경 (Client => ClientConn)
